requirements.txt
requests>=2.28.0
beautifulsoup4>=4.11.0
PyYAML>=6.0
Jinja2>=3.1.0
